[BUG-1] fix notification crash
