+++ app/src/main/java/com/fakhry/lifelog/utils/AlarmReceiver.kt
@@ -1,81 +1,21 @@
 package com.fakhry.lifelog.utils
 
 import android.app.AlarmManager
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.media.RingtoneManager
-import android.os.Build
 import android.widget.Toast
-import androidx.core.app.NotificationCompat
-import androidx.core.content.ContextCompat
 import com.fakhry.lifelog.R
-import com.fakhry.lifelog.ui.activities.main.MainActivity
 import java.text.ParseException
 import java.text.SimpleDateFormat
-import java.util.*
+import java.util.Calendar
+import java.util.Locale
 
 class AlarmReceiver : BroadcastReceiver() {
 
-    companion object {
-        const val TYPE_ONE_TIME = "OneTimeAlarm"
-        const val TYPE_REPEATING = "RepeatingAlarm"
-        const val EXTRA_MESSAGE = "message"
-
-        private const val ID_ONETIME = 100
-        private const val ID_REPEATING = 101
-
-        private const val TIME_FORMAT = "HH:mm"
-
-        private const val CHANNEL_ID = "Channel_1"
-        private const val  CHANNEL_NAME = "AlarmManager channel"
-    }
-
     override fun onReceive(context: Context, intent: Intent) {
-        showAlarmNotification(context)
-    }
-
-    private fun showAlarmNotification(context: Context) {
-
-        val title = context.getString(R.string.notification_title)
-        val message = context.getString(R.string.notification_message)
-
-        val notificationManagerCompat =
-            context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        val alarmSound = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
-
-        val intent = Intent(context, MainActivity::class.java)
-        val pendingIntent = PendingIntent.getActivity(context, 0, intent, 0)
-
-        val builder = NotificationCompat.Builder(context, CHANNEL_ID)
-            .setContentIntent(pendingIntent)
-            .setSmallIcon(R.mipmap.ic_launcher)
-            .setContentTitle(title)
-            .setContentText(message)
-            .setColor(ContextCompat.getColor(context, android.R.color.transparent))
-            .setVibrate(longArrayOf(1000, 1000, 1000, 1000, 1000))
-            .setSound(alarmSound)
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channel = NotificationChannel(
-                CHANNEL_ID,
-                CHANNEL_NAME,
-                NotificationManager.IMPORTANCE_DEFAULT
-            )
-
-            channel.enableVibration(true)
-            channel.vibrationPattern = longArrayOf(1000, 1000, 1000, 1000, 1000)
-
-            builder.setChannelId(CHANNEL_ID)
-
-            notificationManagerCompat.createNotificationChannel(channel)
-        }
-
-        val notification = builder.build()
-        notificationManagerCompat.notify(ID_REPEATING, notification)
+        showReminderNotification(context)
     }
 
     fun setRepeatingAlarm(context: Context, time: String) {
@@ -87,34 +27,48 @@ class AlarmReceiver : BroadcastReceiver() {
         intent.putExtra(EXTRA_MESSAGE, message)
         val timeArray = time.split(":".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()
 
-        val calendar = Calendar.getInstance().also {
-            it.set(Calendar.HOUR_OF_DAY, Integer.parseInt(timeArray[0]))
-            it.set(Calendar.MINUTE, Integer.parseInt(timeArray[1]))
-            it.set(Calendar.SECOND, 0)
+        val calendar = Calendar.getInstance().apply {
+            set(Calendar.HOUR_OF_DAY, timeArray[0].toInt())
+            set(Calendar.MINUTE, timeArray[1].toInt())
+            set(Calendar.SECOND, 0)
         }
 
-        val pendingIntentBroadcast = PendingIntent.getBroadcast(context, ID_REPEATING, intent, 0)
+        val pendingIntentBroadcast = PendingIntent.getBroadcast(
+            context,
+            ID_REPEATING,
+            intent,
+            PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE // Fixed flags
+        )
+
         alarmManager.setInexactRepeating(
             AlarmManager.RTC_WAKEUP,
             calendar.timeInMillis,
             AlarmManager.INTERVAL_DAY,
             pendingIntentBroadcast
         )
-        Toast.makeText(context, context.getString(R.string.reminder_turned_on, time), Toast.LENGTH_SHORT).show()
-    }
 
+        Toast.makeText(
+            context, context.getString(R.string.reminder_turned_on, time), Toast.LENGTH_SHORT
+        ).show()
+    }
 
     fun cancelAlarm(context: Context, type: String) {
         val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
         val intent = Intent(context, AlarmReceiver::class.java)
         val requestCode =
             if (type.equals(TYPE_ONE_TIME, ignoreCase = true)) ID_ONETIME else ID_REPEATING
-        val pendingIntent = PendingIntent.getBroadcast(context, requestCode, intent, 0)
-        pendingIntent.cancel()
+        val pendingIntent = PendingIntent.getBroadcast(
+            context,
+            requestCode,
+            intent,
+            PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE // Fixed flags
+        )
 
+        pendingIntent.cancel()
         alarmManager.cancel(pendingIntent)
 
-        Toast.makeText(context, context.getString(R.string.reminder_turned_off), Toast.LENGTH_SHORT).show()
+        Toast.makeText(context, context.getString(R.string.reminder_turned_off), Toast.LENGTH_SHORT)
+            .show()
     }
 
     private fun isDateInvalid(date: String): Boolean {
@@ -126,5 +80,13 @@ class AlarmReceiver : BroadcastReceiver() {
         } catch (e: ParseException) {
             true
         }
+    }
+
+    companion object {
+        const val EXTRA_MESSAGE = "message"
+
+        private const val ID_ONETIME = 100
+
+        private const val TIME_FORMAT = "HH:mm"
     }
 }
