+++ app/src/main/java/com/fakhry/lifelog/utils/BootReceiver.kt
@@ -1,0 +1,13 @@
+package com.fakhry.lifelog.utils
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+
+class BootReceiver : BroadcastReceiver() {
+    override fun onReceive(context: Context, intent: Intent) {
+        if (Intent.ACTION_BOOT_COMPLETED == intent.action) {
+            showReminderNotification(context)
+        }
+    }
+}
